TFD-1575 Refactoring on constraints model
